requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-bigquery==3.12.0
google-cloud-resource-manager==1.10.4
google-auth==2.23.4
PyJWT==2.8.0
bcrypt==4.1.1
duckdb==0.9.2
pydantic[email]==2.5.0
python-dotenv==1.0.0
structlog==23.2.0
httpx==0.25.2
pandas>=2.1.0
numpy>=1.25.0
